+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixes
 
 - Make sure pipeline runs on multiple samples [#77](https://github.com/nf-core/scrnaseq/pull/77)
+- Fix issue where STARsolo always uses 10XV2 chemistry [#60](https://github.com/nf-core/scrnaseq/issues/60)
 
 ## v1.1.0 - 2021-03-24 "Olive Mercury Corgi"